pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit>=1.0.23",
    "livekit-agents[azure,deepgram,elevenlabs,openai,sarvam,silero,turn-detector]~=1.3",
    "livekit-plugins-assemblyai>=1.3.10",
    "livekit-plugins-cartesia>=1.3.10",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.3.10",
    "livekit-plugins-silero>=1.3.10",
    "livekit-plugins-turn-detector>=1.3.10",
    "python-dotenv~=1.0",
]
